# feat: add support for multiple root directories under a single prefix