--- a/example/public2/test.css
+++ b/example/public2/test.css
@@ -1,0 +1,4 @@
+body {
+  background-color: black;
+  color: white;
+}