--- a/example/public2/test.html
+++ b/example/public2/test.html
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <link rel="stylesheet" type="text/css" href="test.css"></link>
+  </head>
+  <body>
+    <h1>Test 2</h1>
+  </body>
+</html>