--- a/example/server-compress.js
+++ b/example/server-compress.js
@@ -4,9 +4,12 @@ const path = require('path')
 const fastify = require('fastify')({ logger: { level: 'trace' } })
 
 fastify
-  // compress everything
+  // Compress everything.
   .register(require('fastify-compress'), { threshold: 0 })
-  .register(require('../'), { root: path.join(__dirname, '/public') })
+  .register(require('../'), {
+    // An absolute path containing static files to serve.
+    root: path.join(__dirname, '/public')
+  })
   .listen(3000, err => {
     if (err) throw err
   })