--- a/example/server-multi-dir-list.js
+++ b/example/server-multi-dir-list.js
@@ -31,7 +31,6 @@ fastify
   .register(require('..'), {
     root: [path.join(__dirname, '/public'), path.join(__dirname, '/public2')],
     prefixAvoidTrailingSlash: true,
-    wildcard: '**/*.css',
     list: {
       format: 'html',
       names: ['index', 'index.html', 'index.htm', '/'],