--- a/example/server-multi-dir-list.js
+++ b/example/server-multi-dir-list.js
@@ -5,36 +5,42 @@ const Handlebars = require('handlebars')
 
 const fastify = require('fastify')({ logger: { level: 'trace' } })
 
+// Handlebar template for listing files and directories.
 const template = `
-<html><body>
+<html>
+  <body>
+    dirs
+  <ul>
+    {{#dirs}}
+      <li><a href="{{href}}">{{name}}</a></li>
+    {{/dirs}}
+  </ul>
 
-dirs
+  list
 
-<ul>
-{{#dirs}}
-<li><a href="{{href}}">{{name}}</a></li>
-{{/dirs}}
-</ul>
-
-list
-
-<ul>
-{{#files}}
-<li><a href="{{href}}" target="_blank">{{name}}</a></li>
-{{/files}}
-</ul>
-</body></html>
+  <ul>
+    {{#files}}
+      <li><a href="{{href}}" target="_blank">{{name}}</a></li>
+    {{/files}}
+  </ul>
+  </body>
+</html>
 `
 const handlebarTemplate = Handlebars.compile(template)
 
 fastify
   .register(require('..'), {
+    // Array of absolute paths containing static files to serve
     root: [path.join(__dirname, '/public'), path.join(__dirname, '/public2')],
+    // Do not append a trailing slash to prefixes
     prefixAvoidTrailingSlash: true,
-    wildcard: '**/*.css',
+    // Return a directory listing with a handlebar template
     list: {
+      // html or json response? html requires a render method.
       format: 'html',
+      // A list of filenames that trigger a directory list response.
       names: ['index', 'index.html', 'index.htm', '/'],
+      // You can provide your own render method as needed.
       render: (dirs, files) => handlebarTemplate({ dirs, files })
     }
   })