--- a/example/server.js
+++ b/example/server.js
@@ -4,7 +4,7 @@ const path = require('path')
 const fastify = require('fastify')({ logger: { level: 'trace' } })
 
 fastify
-  .register(require('../'), { prefix: '/static/', root: [path.join(__dirname, '/public'), path.join(__dirname, '/public2')] })
+  .register(require('../'), { root: path.join(__dirname, '/public') })
   .listen(3000, err => {
     if (err) throw err
   })