--- a/test/static2/bar.html
+++ b/test/static2/bar.html
@@ -1,0 +1,3 @@
+<html>
+  <body>bar</body>
+</html>