--- a/test/static2/index.html
+++ b/test/static2/index.html
@@ -1,0 +1,3 @@
+<html>
+  <body>index2</body>
+</html>