--- a/test/types/index.ts
+++ b/test/types/index.ts
@@ -53,3 +53,4 @@ multiRootAppWithImplicitHttp
       reply.sendFile('some-file-name')
     })
   })
+