… package.json
… test/static.test.js