<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

This Pull Request aims to add support for multiple `root` directories under a single prefix as per feature suggestion #144. 

Since the `send` module used only accepts a single root path, I have implemented a recursive search approach using the stream error handler. The logic works as follows:

- We take the first path in the root path array and instantiate an instance of `send`.
- If we find the matching file on that path we return it.
- If we find a matching directory on that path we return a listing or index file.
- If we do not find a match we call `pumpSendToReply` again with the same parameters but increment the array offset.
- If we exhaust the root path array we return 404.
 
This approach allows us to return as soon as we have found a matching file/directory without considering all paths and also solves the problem of multiple files with the same name. The first match wins.