@nigelhanlon 2021-01-08
Ran some rough benchmarks to make sure these changes didn't cause any major regressions in performance. 

My machine is a 2017 MacBook Pro, 3.1 GHz Dual-Core Intel Core i5, 8GB RAM.

## Master branch for comparison

On master, my laptop manages roughly 3350 req/seq.  

```
autocannon -c 100 -d 40 -p 10 localhost:3000/index.html
Running 40s test @ http://localhost:3000/index.html
100 connections with 10 pipelining factor

Stat         Avg     Stdev  Max
Latency (ms) 29.6    87.26  942.82
Req/Sec      3348.63 681.4  4077
Bytes/Sec    1.71 MB 348 kB 2.08 MB

134k requests in 40s, 68.2 MB read
```

## Array of root paths, serving from first path

```
autocannon -c 100 -d 40 -p 10 localhost:3000/index.html
Running 40s test @ http://localhost:3000/index.html
100 connections with 10 pipelining factor

Stat         Avg     Stdev  Max
Latency (ms) 28.74   81.73  823.3
Req/Sec      3449.33 479.68 4118
Bytes/Sec    1.76 MB 238 kB 2.1 MB

138k requests in 40s, 70.2 MB read
```

## Array of root paths, serving from second path

Performance is slower since we take a time penalty when we don't find the file in the first path. 

This penalty should increase with every miss.

```
autocannon -c 100 -d 40 -p 10 localhost:3000/test.html
Running 40s test @ http://localhost:3000/test.html
100 connections with 10 pipelining factor

Stat         Avg     Stdev  Max
Latency (ms) 32.06   95.35  1232.6
Req/Sec      3098.78 681.15 3964
Bytes/Sec    1.34 MB 293 kB 1.71 MB

124k requests in 40s, 53.5 MB read
```