@nigelhanlon 2021-01-08
> is this still draft?

This should be ready for final review. Happy to make any changes as needed or suggested.